Paramiko fix for looking for keys
Connecting to Router OS to obtain a configuration via SSH would give an error stating wrong credentials. Credentials are fine, on the router the error log shows:

"ssh,error expected msg: 50 got: 5"

The solution is to not look for local keys, maybe this is not the best practice, but since we already add keys automatically I don't think would be a bigger issue.